requirements.txt
jieba>=0.42.1
requests>=2.31.0
beautifulsoup4>=4.12.0
python-docx>=1.1.0
